+++ esw-gateway/esw-gateway-api/src/main/scala/esw/gateway/api/EventApi.scala
@@ -68,4 +68,17 @@ trait EventApi {
    *         an [[msocket.api.Subscription]] which can be used to unsubscribe from all the Event Keys which were subscribed to
    */
   def pSubscribe(subsystem: Subsystem, maxFrequency: Option[Int] = None, pattern: String = "*"): Source[Event, Subscription]
+
+  /**
+   * Subscribe to all the observe events
+   *
+   * At the time of invocation, in case the underlying server is not available, [[csw.event.api.exceptions.EventServerNotAvailable]] exception is thrown
+   * and the subscription is stopped after logging appropriately. [[csw.event.api.scaladsl.EventSubscription!.ready]] method can be used to determine this
+   * state. In all other cases of exception, the subscription resumes to receive remaining elements.
+   *
+   * @return a [[akka.stream.scaladsl.Source]] of [[csw.params.events.Event]]. The materialized value of the source provides
+   *         an [[csw.event.api.scaladsl.EventSubscription]] which can be used to unsubscribe from observe events subscription
+   */
+  def subscribeObserveEvents(maxFrequency: Option[Int] = None): Source[Event, Subscription]
+
 }